refactor(api): namespace v0 routes and schemas
